Updated docs; restructured repo
