+++ DELETEweb_client/Dockerfile
@@ -1,4 +1,5 @@
-FROM python:3.10
+FROM tiangolo/uwsgi-nginx-flask:python3.10
+# See https://hub.docker.com/r/tiangolo/uwsgi-nginx-flask for info on this image
 
 ARG BRANCH=main
 
@@ -10,8 +11,6 @@ LABEL org.opencontainers.image.licenses=MIT
 RUN apt update ; \
     apt-get -y install nano
 
-WORKDIR /app
-
 COPY ./ /app
 
 RUN pip install --no-cache-dir --upgrade -r /app/requirements.txt
@@ -19,5 +18,3 @@ RUN pip install --no-cache-dir --upgrade -r /app/requirements.txt
 # TODO allow installing from a local branch
 RUN --mount=type=secret,id=git_secrets . /run/secrets/git_secrets \
     && pip install --no-cache-dir git+https://${GIT_USERNAME}:${GIT_PAT}@github.com/ukatc/AtLAST_sensitivity_calculator.git@${BRANCH}#egg=atlast_sc
-
-CMD ["python", "-m", "main"]
