+++ DELETEweb_client/requirements.txt
@@ -1,0 +1,2 @@
+Flask==2.2.*
+flask-wtf==1.0.*
