+++ DELETEweb_client/secrets/.env-copy
@@ -1,0 +1,3 @@
+GIT_USERNAME=fkirton
+GIT_PAT=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
+GIT_CR_PAT=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
