+++ DELETEweb_client/static/css/custom-style.css
