+++ DELETEweb_client/static/js/CalculatorFunctions.js
