+++ DELETEweb_client/static/js/CalculatorFunctions.ts
