+++ DELETEweb_client/static/resources/eu_flag.png
