+++ DELETEweb_client/static/resources/favicon.ico
