+++ environment.yml
@@ -5,8 +5,6 @@ channels:
 dependencies:
   - python=3.10
   - astropy=5.1
-  - Flask=2.2
-  - flask-wtf=1.0
   - numpy=1.23
   - pytest=7.2
   - pytest-mock=3.10
