+++ web_client/main.py
@@ -1,99 +1,78 @@
-import logging
+import uvicorn
+from fastapi import FastAPI, HTTPException, Request
+from fastapi.responses import HTMLResponse, JSONResponse
+from fastapi.templating import Jinja2Templates
+from fastapi.staticfiles import StaticFiles
+from schemas import APIUserInput
+import crud
+import context_processors as cp
 
-from flask import Flask, render_template, request, jsonify
-
-import astropy.units as u
-from atlast_sc.calculator import Calculator
-
-app = Flask(__name__)
-
-handler = logging.StreamHandler()
-app.logger.addHandler(handler)
-app.logger.propagate = False
-app.logger.level = logging.DEBUG
+app = FastAPI()
 
 
-@app.route('/', methods=('GET', 'POST'))
-def template():
-    return render_template('SensitivityCalculator.html')
+templates = Jinja2Templates(directory="templates",
+                            context_processors=[cp.invalid_message_processor,
+                                                cp.default_values_processor,
+                                                cp.default_units_processor,
+                                                cp.allowed_range_processor,
+                                                ])
+
+app.mount("/static", StaticFiles(directory="static"), name="static")
+app.mount("/scripts", StaticFiles(directory="scripts"), name="scripts")
 
 
-@app.route('/hello')
-def hello():
-    return "Hello2 World"
+@app.get("/", response_class=HTMLResponse)
+async def sensitivity_calculator(request: Request):
+
+    return templates.TemplateResponse("sensitivity_calculator.html",
+                                      {"request": request,
+                                       "params_vals_units":
+                                           crud.get_param_values_units()})
 
 
-@app.route('/documentation')
-def docs():
-    return "To build the html version of the documentation, " \
-           "start from the main package directory " \
-           "and type ``cd docs; make html'. Read the documentation" \
-           " by pointing your browser at " \
-           "``{}/docs/build/html/index.html".format(app.root_path)
+@app.post("/v1/sensitivity")
+async def sensitivity(api_user_input: APIUserInput):
+
+    user_input = _unpack_api_user_input(api_user_input)
+
+    try:
+        return crud.do_calculation(user_input, "sensitivity")
+    except crud.UserInputError as e:
+        raise HTTPException(status_code=400, detail=e.message)
 
 
-@app.route('/v1/sensitivity')
-def sensitivity():
-    app.logger.debug('sensitivity1')
-    print('received request', request)
+@app.post("/v1/integration-time")
+async def t_int(api_user_input: APIUserInput):
 
-    inputs = {}
-    inputs['bandwidth'] = {
-        'value': float(request.args.get('bandwidth')),
-        'unit': 'GHz'
+    user_input = _unpack_api_user_input(api_user_input)
+
+    try:
+        return crud.do_calculation(user_input, "integration_time")
+    except crud.UserInputError as e:
+        raise HTTPException(status_code=400, detail=e.message)
+
+
+@app.post("/v1/test")
+async def test(test_input: dict):
+    return test_input
+
+
+@app.get("/v1/param-values-units")
+async def param_values_units():
+    return JSONResponse(content=crud.get_param_values_units())
+
+
+def _unpack_api_user_input(api_user_input):
+    return {
+        "t_int": api_user_input.t_int,
+        "sensitivity": api_user_input.sensitivity,
+        "bandwidth": api_user_input.bandwidth,
+        "obs_freq": api_user_input.obs_freq,
+        "n_pol": api_user_input.n_pol,
+        "weather": api_user_input.weather,
+        "elevation": api_user_input.elevation,
     }
-    inputs['obs_freq'] = {
-        'value': float(request.args.get('obs_freq')),
-        'unit': 'GHz'
-    }
-    inputs['n_pol'] = {
-        'value': int(request.args.get('npol')),
-        'unit': 'none'
-    }
-    inputs['weather'] = {
-        'value': float(request.args.get('pwv')),
-        'unit': 'none'
-    }
-    inputs['elevation'] = {
-        'value': float(request.args.get('elevation')),
-        'unit': 'deg'
-    }
 
-    if 'integration_time' in request.args:
-        inputs['t_int'] = {
-            'value': float(request.args.get('integration_time')),
-            'unit': 's'
-        }
-    if 'sensitivity' in request.args:
-        inputs['sensitivity'] = {
-            'value': float(request.args.get('sensitivity')),
-            'unit': 'mJy'
-        }
 
-    app.logger.debug(inputs)
-
-    # config = Config(inputs)
-    calculator = Calculator(inputs)
-
-    result_dict = {}
-    # TODO: the requirement that exactly one of sensitivity and int time
-    #  should have a value is causing and error.
-    #       Remove this requirement. It's not adding anything useful.
-    if 'integration_time' in request.args:
-        result = \
-            calculator.calculate_sensitivity(calculator.t_int,
-                                             update_calculator=False).to(u.mJy)
-        app.logger.debug('calculator.calculate_sensitivity')
-        app.logger.debug(result)
-        result_dict["sensitivity"] = f"{result:0.03f}"
-    elif 'sensitivity' in request.args:
-        result = calculator.calculate_t_integration(calculator.sensitivity,
-                                                    update_calculator=False)
-        app.logger.debug('calculator.calculate_t_integration')
-        app.logger.debug(result)
-        result_dict["integration_time"] = f"{result:0.03f}"
-
-    app.logger.debug(result_dict)
-
-    result = jsonify(result_dict)
-    return result
+if __name__ == "__main__":
+    uvicorn.run(app, host="0.0.0.0", port=8000)
