+++ web_client/requirements.txt
@@ -1,2 +1,4 @@
-Flask==2.2.*
-flask-wtf==1.0.*
+fastapi==0.94.*
+uvicorn==0.21.*
+jinja2==3.1.*
+aiofiles==22.1.*
