+++ web_client/scripts/calculator_ui.js
