+++ web_client/scripts/rest_calls.js
